test(google-meet): align CLI split routing
