+++ extensions/google-meet/src/cli-artifacts.test.ts
@@ -1,4 +1,4 @@
-import { mkdtempSync, readFileSync, rmSync } from "node:fs";
+import { existsSync, mkdtempSync, readFileSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import JSZip from "jszip";
@@ -22,57 +22,6 @@ describe("google-meet CLI", () => {
   afterAll(() => {
     vi.doUnmock("openclaw/plugin-sdk/ssrf-runtime");
     vi.resetModules();
-  });
-
-  it("prints setup checks as text and JSON", async () => {
-    {
-      const stdout = captureStdout();
-      try {
-        await setupCli({
-          runtime: {
-            setupStatus: async () => ({
-              ok: true,
-              checks: [
-                {
-                  id: "audio-bridge",
-                  ok: true,
-                  message: "Chrome command-pair talk-back audio bridge configured (pcm16-24khz)",
-                },
-              ],
-            }),
-          },
-        }).parseAsync(["googlemeet", "setup"], { from: "user" });
-        expect(stdout.output()).toContain("Google Meet setup: OK");
-        expect(stdout.output()).toContain(
-          "[ok] audio-bridge: Chrome command-pair talk-back audio bridge configured (pcm16-24khz)",
-        );
-        expect(stdout.output()).not.toContain('"checks"');
-      } finally {
-        stdout.restore();
-      }
-    }
-
-    {
-      const stdout = captureStdout();
-      try {
-        await setupCli({
-          runtime: {
-            setupStatus: async () => ({
-              ok: false,
-              checks: [{ id: "twilio-voice-call-plugin", ok: false, message: "missing" }],
-            }),
-          },
-        }).parseAsync(["googlemeet", "setup", "--json"], { from: "user" });
-        const payload = parseStdoutJson(stdout);
-        expectFields(payload, { ok: false });
-        expectFields(firstRecord(payload.checks), {
-          id: "twilio-voice-call-plugin",
-          ok: false,
-        });
-      } finally {
-        stdout.restore();
-      }
-    }
   });
 
   it("prints artifacts and attendance output", async () => {
@@ -579,6 +528,59 @@ describe("google-meet CLI", () => {
     } finally {
       stdout.restore();
       rmSync(tempDir, { recursive: true, force: true });
+    }
+  });
+
+  it("prints a dry-run export manifest without writing files", async () => {
+    stubMeetArtifactsApi();
+    const stdout = captureStdout();
+    const parentDir = mkdtempSync(path.join(tmpdir(), "openclaw-google-meet-export-dry-run-"));
+    const outputDir = path.join(parentDir, "bundle");
+
+    try {
+      await setupCli({}).parseAsync(
+        [
+          "googlemeet",
+          "export",
+          "--access-token",
+          "token",
+          "--expires-at",
+          String(Date.now() + 120_000),
+          "--conference-record",
+          "rec-1",
+          "--include-doc-bodies",
+          "--output",
+          outputDir,
+          "--dry-run",
+        ],
+        { from: "user" },
+      );
+      const payload = JSON.parse(stdout.output());
+      expectFields(payload, {
+        dryRun: true,
+        tokenSource: "cached-access-token",
+      });
+      expectFields(payload.manifest.request, {
+        conferenceRecord: "rec-1",
+        includeDocumentBodies: true,
+      });
+      expectFields(payload.manifest.counts, {
+        attendanceRows: 1,
+        transcriptEntries: 1,
+        warnings: 0,
+      });
+      expect(payload.manifest.files).toEqual([
+        "summary.md",
+        "attendance.csv",
+        "transcript.md",
+        "artifacts.json",
+        "attendance.json",
+        "manifest.json",
+      ]);
+      expect(existsSync(outputDir)).toBe(false);
+    } finally {
+      stdout.restore();
+      rmSync(parentDir, { recursive: true, force: true });
     }
   });
 });
